Move the disc-ral dialect into its own directory structure.
The mlir-hlo directory has grown to a size where some extra structure seems needed.

PiperOrigin-RevId: 407548213
