+++ tensorflow/compiler/mlir/hlo/BUILD
@@ -677,13 +677,45 @@ gentbl_cc_library(
                 "-gen-pass-decls",
                 "-name=RAL",
             ],
-            "include/mlir-hlo/Dialect/mhlo/transforms/disc_ral_passes.h.inc",
+            "include/mlir-hlo/Dialect/disc-ral/transforms/disc_ral_passes.h.inc",
         ),
     ],
     tblgen = "@llvm-project//mlir:mlir-tblgen",
-    td_file = "include/mlir-hlo/Dialect/mhlo/transforms/disc_ral_passes.td",
+    td_file = "include/mlir-hlo/Dialect/disc-ral/transforms/disc_ral_passes.td",
     deps = [
         "@llvm-project//mlir:PassBaseTdFiles",
+    ],
+)
+
+td_library(
+    name = "disc_ral_ops_td_files",
+    srcs = glob(["include/mlir-hlo/Dialect/disc-ral/IR/*.td"]),
+    compatible_with = get_compatible_with_cloud(),
+    includes = ["include"],
+    deps = [
+        "@llvm-project//mlir:ControlFlowInterfacesTdFiles",
+        "@llvm-project//mlir:CopyOpInterfaceTdFiles",
+        "@llvm-project//mlir:InferTypeOpInterfaceTdFiles",
+        "@llvm-project//mlir:LoopLikeInterfaceTdFiles",
+        "@llvm-project//mlir:MemRefOpsTdFiles",
+        "@llvm-project//mlir:OpBaseTdFiles",
+        "@llvm-project//mlir:ShapeOpsTdFiles",
+        "@llvm-project//mlir:SideEffectTdFiles",
+        "@llvm-project//mlir:ViewLikeInterfaceTdFiles",
+    ],
+)
+
+cc_library(
+    name = "disc_ral_pass_details",
+    hdrs = [
+        "include/mlir-hlo/Dialect/disc-ral/transforms/PassDetail.h",
+    ],
+    visibility = [
+        "//visibility:private",  # This target is a private detail of pass implementations
+    ],
+    deps = [
+        ":DiscRalPassIncGen",
+        "@llvm-project//mlir:Pass",
     ],
 )
 
@@ -694,27 +726,27 @@ gentbl_cc_library(
     tbl_outs = [
         (
             ["-gen-op-decls"],
-            "include/mlir-hlo/Dialect/mhlo/IR/disc_ral_ops.h.inc",
+            "include/mlir-hlo/Dialect/disc-ral/IR/disc_ral_ops.h.inc",
         ),
         (
             ["-gen-op-defs"],
-            "include/mlir-hlo/Dialect/mhlo/IR/disc_ral_ops.cc.inc",
+            "include/mlir-hlo/Dialect/disc-ral/IR/disc_ral_ops.cc.inc",
         ),
     ],
     tblgen = "@llvm-project//mlir:mlir-tblgen",
-    td_file = "include/mlir-hlo/Dialect/mhlo/IR/disc_ral_ops.td",
-    deps = [":hlo_ops_td_files"],
+    td_file = "include/mlir-hlo/Dialect/disc-ral/IR/disc_ral_ops.td",
+    deps = [":disc_ral_ops_td_files"],
 )
 
 cc_library(
     name = "disc_ral",
     srcs = [
-        "include/mlir-hlo/Dialect/mhlo/IR/disc_ral_ops.cc.inc",
-        "include/mlir-hlo/Dialect/mhlo/IR/disc_ral_ops.h.inc",
-        "lib/Dialect/mhlo/IR/disc_ral_ops.cc",
+        "include/mlir-hlo/Dialect/disc-ral/IR/disc_ral_ops.cc.inc",
+        "include/mlir-hlo/Dialect/disc-ral/IR/disc_ral_ops.h.inc",
+        "lib/Dialect/disc-ral/IR/disc_ral_ops.cc",
     ],
     hdrs = [
-        "include/mlir-hlo/Dialect/mhlo/IR/disc_ral_ops.h",
+        "include/mlir-hlo/Dialect/disc-ral/IR/disc_ral_ops.h",
     ],
     includes = ["include"],
     deps = [
@@ -739,11 +771,11 @@ cc_library(
 
 cc_library(
     name = "ral_inject_execution_context",
-    srcs = ["lib/Dialect/mhlo/transforms/ral_inject_execution_context.cc"],
-    hdrs = ["include/mlir-hlo/Dialect/mhlo/transforms/passes.h"],
+    srcs = ["lib/Dialect/disc-ral/transforms/ral_inject_execution_context.cc"],
+    hdrs = ["include/mlir-hlo/Dialect/disc-ral/transforms/passes.h"],
     deps = [
         ":disc_ral",
-        ":pass_details",
+        ":disc_ral_pass_details",
         "@llvm-project//llvm:Support",
         "@llvm-project//mlir:IR",
         "@llvm-project//mlir:Pass",
@@ -757,13 +789,13 @@ cc_library(
 
 cc_library(
     name = "ral_lower_to_library_call",
-    srcs = ["lib/Dialect/mhlo/transforms/ral_lower_to_library_call.cc"],
+    srcs = ["lib/Dialect/disc-ral/transforms/ral_lower_to_library_call.cc"],
     hdrs = [
-        "include/mlir-hlo/Dialect/mhlo/transforms/passes.h",
+        "include/mlir-hlo/Dialect/disc-ral/transforms/passes.h",
     ],
     deps = [
         ":disc_ral",
-        ":pass_details",
+        ":disc_ral_pass_details",
         "@llvm-project//llvm:Support",
         "@llvm-project//llvm:TransformUtils",
         "@llvm-project//mlir:IR",
@@ -779,14 +811,14 @@ cc_library(
 
 cc_library(
     name = "ral_legalize_to_llvm",
-    srcs = ["lib/Dialect/mhlo/transforms/ral_legalize_to_llvm.cc"],
+    srcs = ["lib/Dialect/disc-ral/transforms/ral_legalize_to_llvm.cc"],
     hdrs = [
-        "include/mlir-hlo/Dialect/mhlo/transforms/passes.h",
+        "include/mlir-hlo/Dialect/disc-ral/transforms/passes.h",
+        "include/mlir-hlo/Dialect/disc-ral/transforms/rewriters.h",
     ],
     deps = [
         ":disc_ral",
-        ":mhlo_to_mhlo_lowering_patterns",
-        ":pass_details",
+        ":disc_ral_pass_details",
         "@llvm-project//llvm:Support",
         "@llvm-project//llvm:TransformUtils",
         "@llvm-project//mlir:Affine",
@@ -1611,7 +1643,6 @@ cc_library(
         "//visibility:private",  # This target is a private detail of pass implementations
     ],
     deps = [
-        ":DiscRalPassIncGen",
         ":LmhloPassIncGen",
         ":MhloPassIncGen",
         "@llvm-project//mlir:Pass",
@@ -1660,6 +1691,7 @@ cc_library(
 cc_library(
     name = "all_passes",
     hdrs = [
+        "include/mlir-hlo/Dialect/disc-ral/transforms/register_passes.h",
         "include/mlir-hlo/Dialect/mhlo/transforms/register_passes.h",
         "include/mlir-hlo/Transforms/register_passes.h",
     ],
