+++ tensorflow/compiler/mlir/hlo/include/mlir-hlo/Dialect/disc-ral/transforms/CMakeLists.txt
@@ -14,10 +14,6 @@
 # limitations under the License.
 #
 
-set(LLVM_TARGET_DEFINITIONS mhlo_passes.td)
-mlir_tablegen(mhlo_passes.h.inc -gen-pass-decls -name MHLO)
-add_public_tablegen_target(MLIRMhloPassIncGen)
-
-set(LLVM_TARGET_DEFINITIONS lmhlo_passes.td)
-mlir_tablegen(lmhlo_passes.h.inc -gen-pass-decls -name LMHLO)
-add_public_tablegen_target(MLIRLmhloPassIncGen)
+set(LLVM_TARGET_DEFINITIONS disc_ral_passes.td)
+mlir_tablegen(disc_ral_passes.h.inc -gen-pass-decls -name RAL)
+add_public_tablegen_target(MLIRDiscRalPassIncGen)
