+++ tensorflow/compiler/mlir/hlo/include/mlir-hlo/Dialect/disc-ral/transforms/PassDetail.h
@@ -13,33 +13,18 @@ See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
 
-#ifndef TENSORFLOW_COMPILER_MLIR_HLO_INCLUDE_MLIR_HLO_DIALECT_MHLO_TRANSFORMS_PASSDETAIL_H_
-#define TENSORFLOW_COMPILER_MLIR_HLO_INCLUDE_MLIR_HLO_DIALECT_MHLO_TRANSFORMS_PASSDETAIL_H_
+#ifndef TENSORFLOW_COMPILER_MLIR_HLO_INCLUDE_MLIR_HLO_DIALECT_DISC_RAL_TRANSFORMS_PASSDETAIL_H_
+#define TENSORFLOW_COMPILER_MLIR_HLO_INCLUDE_MLIR_HLO_DIALECT_DISC_RAL_TRANSFORMS_PASSDETAIL_H_
 
 #include "mlir/Pass/Pass.h"
 
 namespace mlir {
-namespace scf {
-class SCFDialect;
-}
-namespace memref {
-class MemRefDialect;
-}
-
-namespace mhlo {
-class MhloDialect;
-#define GEN_PASS_CLASSES
-#include "mlir-hlo/Dialect/mhlo/transforms/mhlo_passes.h.inc"
-
-}  // end namespace mhlo
-
-namespace lmhlo {
+namespace disc_ral {
 
 #define GEN_PASS_CLASSES
-#include "mlir-hlo/Dialect/mhlo/transforms/lmhlo_passes.h.inc"
+#include "mlir-hlo/Dialect/disc-ral/transforms/disc_ral_passes.h.inc"
 
-}  // end namespace lmhlo
-
+}  // end namespace disc_ral
 }  // end namespace mlir
 
-#endif  // TENSORFLOW_COMPILER_MLIR_HLO_INCLUDE_MLIR_HLO_DIALECT_MHLO_TRANSFORMS_PASSDETAIL_H_
+#endif  // TENSORFLOW_COMPILER_MLIR_HLO_INCLUDE_MLIR_HLO_DIALECT_DISC_RAL_TRANSFORMS_PASSDETAIL_H_
