+++ tensorflow/compiler/mlir/hlo/include/mlir-hlo/Dialect/disc-ral/transforms/rewriters.h
@@ -1,0 +1,38 @@
+/* Copyright 2019 The TensorFlow Authors. All Rights Reserved.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+==============================================================================*/
+
+#ifndef TENSORFLOW_COMPILER_MLIR_HLO_INCLUDE_MLIR_HLO_DIALECT_DISC_RAL_TRANSFORMS_REWRITERS_H_
+#define TENSORFLOW_COMPILER_MLIR_HLO_INCLUDE_MLIR_HLO_DIALECT_DISC_RAL_TRANSFORMS_REWRITERS_H_
+
+#include <memory>
+
+#include "mlir/Transforms/DialectConversion.h"
+
+namespace mlir {
+
+class LLVMTypeConverter;
+class SymbolTable;
+
+namespace disc_ral {
+
+void populateDiscRalToLLVMConversionPatterns(LLVMTypeConverter *converter,
+                                             SymbolTable *symbol_table,
+                                             RewritePatternSet *patterns);
+
+}  // namespace disc_ral
+
+}  // namespace mlir
+
+#endif  // TENSORFLOW_COMPILER_MLIR_HLO_INCLUDE_MLIR_HLO_DIALECT_DISC_RAL_TRANSFORMS_REWRITERS_H_
