+++ tensorflow/compiler/mlir/hlo/include/mlir-hlo/Dialect/mhlo/transforms/passes.h
@@ -148,19 +148,6 @@ std::unique_ptr<FunctionPass> createInputInlineFusionPass();
 
 }  // namespace lmhlo
 
-namespace disc_ral {
-
-std::unique_ptr<OperationPass<ModuleOp>> createRalInjectExecutionContextPass(
-    const std::string& entry_func_name = "main");
-
-// Lower some specific ops to library calls (modeled by `disc_ral.launch` op).
-std::unique_ptr<mlir::FunctionPass> createRalLowerToLibraryCallPass();
-
-// Lower disc to llvm dialect
-std::unique_ptr<OperationPass<ModuleOp>> createRalToLLVMPass();
-
-}  // namespace disc_ral
-
 }  // namespace mlir
 
 #endif  // TENSORFLOW_COMPILER_MLIR_HLO_INCLUDE_MLIR_HLO_DIALECT_MHLO_TRANSFORMS_PASSES_H_
