+++ tensorflow/compiler/mlir/hlo/include/mlir-hlo/Dialect/mhlo/transforms/register_passes.h
@@ -43,15 +43,4 @@ inline void registerAllLmhloPasses() { registerLMHLOPasses(); }
 }  // namespace lmhlo
 }  // namespace mlir
 
-namespace mlir {
-namespace disc_ral {
-
-#define GEN_PASS_REGISTRATION
-#include "mlir-hlo/Dialect/mhlo/transforms/disc_ral_passes.h.inc"
-
-inline void registerAllDiscRalPasses() { registerRALPasses(); }
-
-}  // namespace disc_ral
-}  // namespace mlir
-
 #endif  // MLIR_HLO_DIALECT_MHLO_TRANSFORMS_REGISTER_PASSES_H_
