+++ tensorflow/compiler/mlir/hlo/lib/Dialect/CMakeLists.txt
@@ -13,4 +13,5 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
+add_subdirectory(disc-ral)
 add_subdirectory(mhlo)
