+++ tensorflow/compiler/mlir/hlo/lib/Dialect/disc-ral/CMakeLists.txt
@@ -13,11 +13,5 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-
-set(LLVM_TARGET_DEFINITIONS mhlo_passes.td)
-mlir_tablegen(mhlo_passes.h.inc -gen-pass-decls -name MHLO)
-add_public_tablegen_target(MLIRMhloPassIncGen)
-
-set(LLVM_TARGET_DEFINITIONS lmhlo_passes.td)
-mlir_tablegen(lmhlo_passes.h.inc -gen-pass-decls -name LMHLO)
-add_public_tablegen_target(MLIRLmhloPassIncGen)
+add_subdirectory(IR)
+add_subdirectory(transforms)
