+++ tensorflow/compiler/mlir/hlo/lib/Dialect/disc-ral/IR/disc_ral_ops.cc
@@ -15,7 +15,7 @@ limitations under the License.
 
 // This file defines the operations used in the DISC RAL dialect.
 
-#include "mlir-hlo/Dialect/mhlo/IR/disc_ral_ops.h"
+#include "mlir-hlo/Dialect/disc-ral/IR/disc_ral_ops.h"
 
 namespace mlir {
 namespace disc_ral {
@@ -33,7 +33,7 @@ RalDialect::RalDialect(MLIRContext* context)
     : Dialect(getDialectNamespace(), context, TypeID::get<RalDialect>()) {
   addOperations<
 #define GET_OP_LIST
-#include "mlir-hlo/Dialect/mhlo/IR/disc_ral_ops.cc.inc"
+#include "mlir-hlo/Dialect/disc-ral/IR/disc_ral_ops.cc.inc"
       >();
   addTypes<RalExecutionContextType>();
   context->loadDialect<memref::MemRefDialect>();
@@ -61,4 +61,4 @@ void RalDialect::printType(Type type, DialectAsmPrinter& os) const {
 }  // namespace mlir
 
 #define GET_OP_CLASSES
-#include "mlir-hlo/Dialect/mhlo/IR/disc_ral_ops.cc.inc"
+#include "mlir-hlo/Dialect/disc-ral/IR/disc_ral_ops.cc.inc"
