+++ tensorflow/compiler/mlir/hlo/tools/mlir-hlo-opt/CMakeLists.txt
@@ -21,6 +21,7 @@ set(LIBS
         MLIROptLib
 
         MhloRegisterDialects
+        AllDiscRalPasses
         AllMhloPasses
         MLIRBufferTransforms
         MLIRHLOAnalysis
@@ -28,6 +29,7 @@ set(LIBS
         )
 add_llvm_executable(mlir-hlo-opt mlir-hlo-opt.cpp
   DEPENDS
+        MLIRDiscRalPassIncGen
         MLIRLmhloPassIncGen
         MLIRMhloPassIncGen
         LMHLOTransformsPassIncGen
